refactor: move shared libs function (getCategory, getPost) to entities folder
