+++ src/entities/category/model/getCategories.ts
